app/requirements.txt
streamlit>=1.28.0
python-docx>=0.8.11
PyYAML>=6.0
pandas>=2.0.0
